hook position closed loop shuffleboard

+++ src/main/java/frc/robot/commands/HookPositionWithMagic.java
@@ -3,7 +3,7 @@ package frc.robot.commands;
 import frc.robot.subsystems.HookSubsystem;
 
 public class HookPositionWithMagic extends HookPosition {
-  public HookPositionWithMagic(HookSubsystem hook, int setpoint) {
+  public HookPositionWithMagic(HookSubsystem hook, double setpoint) {
     super(hook, setpoint);
   }
